Ticket 1423: Pipe needed to changed to size_t to compile in 64 bit GDC

+++ tango/sys/Pipe.d
@@ -76,7 +76,7 @@ class PipeConduit : Device
     /**
      * Returns the buffer size for the PipeConduit.
      */
-    public override uint bufferSize()
+    public override size_t bufferSize()
     {
         return _bufferSize;
     }
